Minor reorganization of async projection code
