+++ src/Marten/Events/Daemon/AgentStatus.cs
@@ -1,0 +1,9 @@
+namespace Marten.Events.Daemon
+{
+    public enum AgentStatus
+    {
+        Running,
+        Stopped,
+        Paused
+    }
+}
